+++ src/collective/documentgenerator/migrations/migrate_to_6.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+
+import logging
+from imio.migrator.migrator import Migrator
+from plone import api
+
+logger = logging.getLogger('collective.documentgenerator')
+
+
+class Migrate_To_6(Migrator):
+
+    def __init__(self, context):
+        Migrator.__init__(self, context)
+        self.catalog = api.portal.get_tool('portal_catalog')
+
+    def run(self):
+        logger.info('Migrating to collective.documentgenerator 6 ...')
+        self.runProfileSteps('collective.documentgenerator', steps=['typeinfo'], profile='install-base')
+        self.finish()
+
+
+def migrate(context):
+    '''
+    '''
+    Migrate_To_6(context).run()